+++ pyproject.toml
@@ -1,0 +1,25 @@
+[project]
+name = "file-mapper"
+version = "0.1.0"
+description = "File mapping utility for BIDS data processing"
+readme = "README.md"
+requires-python = ">=3.11"
+dependencies = [
+    "pyqt5>=5.15.11",
+]
+
+[project.scripts]
+filemapper = "filemapper:main"
+filemapper-gui = "gui:main"
+
+[build-system]
+requires = ["setuptools>=61.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools]
+py-modules = ["filemapper", "gui", "destination_window_gui"]
+
+[dependency-groups]
+dev = [
+    "black>=25.1.0",
+]
